requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
environ==1.0
pytz==2022.7.1
sqlparse==0.4.3
